terminal support on shell server
