+++ Makefile
@@ -14,7 +14,7 @@
 VERSION = 1
 PATCHLEVEL = 5
 SUBLEVEL = 0
-EXTRAVERSION = -rc4
+EXTRAVERSION = -rc5
 NAME =
